\name{hydroBE}
\alias{hydroBE}
\docType{package}

\title{
\packageTitle{hydroBE}
}

\description{
\packageDescription{hydroBE}
}

\details{
The present mingle-mangle of functions was written during our work in
the group of hydrology led by Rolf Weingartner at the University of
Bern. The package is not included in an official R package repository,
however, it should pass the CRAN check.
}

\author{
\packageAuthor{hydroBE}

Maintainer: \packageMaintainer{hydroBE}
}

\keyword{package}
