\name{hydroBE-package}
\alias{hydroBE-package}
\docType{package}

\title{
  \packageTitle{hydroBE}
}

\description{
  \packageDescription{hydroBE}
}

\details{
  The present mingle-mangle of functions was written during our work in
  the group of hydrology led by Rolf Weingartner at the University of
  Bern.
}

\keyword{package}
