\name{hydroBE-package}
\alias{hydroBE-package}
\alias{hydroBE}
\docType{package}
\title{
\packageTitle{hydroBE}
}
\description{
\packageDescription{hydroBE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hydroBE}
\packageIndices{hydroBE}
%% ~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hydroBE} (Maintainer: \packageMaintainer{hydroBE})
}
\references{
%% ~~ Literature or other references for background information ~~
}
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
