\name{hydroBE-package}
\alias{hydroBE-package}
\alias{hydroBE}
\docType{package}

\title{
\packageTitle{hydroBE}
}

\description{
\packageDescription{hydroBE}
}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{hydroBE}
\packageIndices{hydroBE}
}

\author{
\packageAuthor{hydroBE} 
%- (Maintainer: \packageMaintainer{hydroBE})
}

\keyword{ package }